third_party/blink/web_tests/external/wpt/svg/animations/svgangle-animation-unitType-expected.txt

This is a testharness.js-based test.
[FAIL] Test change of unit type for SVGAngle animation.
  assert_equals: expected 4 but got 2
Harness: the test ran to completion.
